Week_01
4) Explain the difference between interfaces and implementation:

Interface: It is set of functionalities that allows the use of system. The methods for interface must be available to implement it. interface is a prototype of the function that can be implemented throughout the system implementation.

Implementation: An implementation implements interface to the system and to create something according to design pattern and specifications provided from interface.

5) Using both visual and written descriptions, think through the interface-implementation of a large scale storage system.   In many systems today, we have the ability to store information from a single application to a variety of storage devices - local storage (hard drive, usb), the cloud and/or some new medium in the future.   How would you design an interface structure such that all of the possible implementations could store data effectively.

The way this can be implemented where the files can be stored on different platforms by giving the options is to give the user the selection of choices where they want the files to be stored it can be one or more.